Move tracing package to internal
